--3pt logic  working

+++ app/backend/main.py
@@ -2863,9 +2863,9 @@ def get_full_game_state(
         "basket_position": basket_pos_py,
         "court_width": game_state.env.court_width,
         "court_height": game_state.env.court_height,
-        "three_point_distance": int(getattr(game_state.env, "three_point_distance", 4)),
+        "three_point_distance": float(getattr(game_state.env, "three_point_distance", 4.0)),
         "three_point_short_distance": (
-            int(getattr(game_state.env, "three_point_short_distance"))
+            float(getattr(game_state.env, "three_point_short_distance"))
             if getattr(game_state.env, "three_point_short_distance", None) is not None
             else None
         ),
